Refactor imports for @hedvig-ui/DateTimePicker, again
