+++ shared/hedvig-ui/DateTimePicker/date-time-picker.stories.tsx
@@ -1,4 +1,4 @@
-import { DateTimePicker } from 'hedvig-ui/date-time-picker'
+import { DateTimePicker } from '@hedvig-ui'
 import React from 'react'
 
 export default {
